+++ packages/cosmwasm/src/index.ts
@@ -25,7 +25,6 @@ export {
   InstantiateOptions,
   InstantiateResult,
   MigrateResult,
-  SigningCallback,
   SigningCosmWasmClient,
   UploadMeta,
   UploadResult,