+++ packages/cosmwasm/src/signingcosmwasmclient.ts
@@ -13,7 +13,6 @@ import {
   PostTxFailure,
   PostTxResult,
   StdFee,
-  StdSignature,
   StdTx,
 } from "@cosmjs/launchpad";
 import { Uint53 } from "@cosmjs/math";
@@ -31,10 +30,9 @@ import {
   MsgUpdateAdmin,
 } from "./msgs";
 
-export interface SigningCallback {
-  (signBytes: Uint8Array): Promise<StdSignature>;
-}
-
+/**
+ * Those fees are used by the higher level methods of SigningCosmWasmClient
+ */
 export interface FeeTable {
   readonly upload: StdFee;
   readonly init: StdFee;