+++ packages/cosmwasm/types/signingcosmwasmclient.d.ts
@@ -1,17 +1,9 @@
-import {
-  BroadcastMode,
-  Coin,
-  Msg,
-  OfflineSigner,
-  PostTxResult,
-  StdFee,
-  StdSignature,
-} from "@cosmjs/launchpad";
+import { BroadcastMode, Coin, Msg, OfflineSigner, PostTxResult, StdFee } from "@cosmjs/launchpad";
 import { Account, CosmWasmClient, GetSequenceResult } from "./cosmwasmclient";
 import { Log } from "./logs";
-export interface SigningCallback {
-  (signBytes: Uint8Array): Promise<StdSignature>;
-}
+/**
+ * Those fees are used by the higher level methods of SigningCosmWasmClient
+ */
 export interface FeeTable {
   readonly upload: StdFee;
   readonly init: StdFee;